# use VertexArrayObject extension when available