--- a/js/data/bucket/circle_bucket.js
+++ b/js/data/bucket/circle_bucket.js
@@ -21,7 +21,7 @@ function CircleBucket() {
 CircleBucket.prototype = util.inherit(Bucket, {});
 
 CircleBucket.prototype.addCircleVertex = function(x, y, extrudeX, extrudeY) {
-    return this.arrays.circleVertex.emplaceBack(
+    return this.arrays.circle.layout.vertex.emplaceBack(
             (x * 2) + ((extrudeX + 1) / 2),
             (y * 2) + ((extrudeY + 1) / 2));
 };
@@ -63,7 +63,7 @@ CircleBucket.prototype.addFeature = function(feature) {
     var globalProperties = {zoom: this.zoom};
     var geometries = loadGeometry(feature);
 
-    var startIndex = this.arrays.circleVertex.length;
+    var startIndex = this.arrays.circle.layout.vertex.length;
 
     for (var j = 0; j < geometries.length; j++) {
         for (var k = 0; k < geometries[j].length; k++) {
@@ -91,11 +91,11 @@ CircleBucket.prototype.addFeature = function(feature) {
             this.addCircleVertex(x, y, -1, 1);
             group.vertexLength += 4;
 
-            this.arrays.circleElement.emplaceBack(index, index + 1, index + 2);
-            this.arrays.circleElement.emplaceBack(index, index + 3, index + 2);
+            this.arrays.circle.layout.element.emplaceBack(index, index + 1, index + 2);
+            this.arrays.circle.layout.element.emplaceBack(index, index + 3, index + 2);
             group.elementLength += 2;
         }
     }
 
-    this.addPaintAttributes('circle', globalProperties, feature.properties, startIndex, this.arrays.circleVertex.length);
+    this.addPaintAttributes('circle', globalProperties, feature.properties, startIndex, this.arrays.circle.layout.vertex.length);
 };