--- a/js/data/bucket/line_bucket.js
+++ b/js/data/bucket/line_bucket.js
@@ -51,7 +51,7 @@ function LineBucket() {
 LineBucket.prototype = util.inherit(Bucket, {});
 
 LineBucket.prototype.addLineVertex = function(point, extrude, tx, ty, dir, linesofar) {
-    return this.arrays.lineVertex.emplaceBack(
+    return this.arrays.line.layout.vertex.emplaceBack(
             // a_pos
             (point.x << 1) | tx,
             (point.y << 1) | ty,
@@ -375,7 +375,7 @@ LineBucket.prototype.addCurrentVertex = function(currentVertex, distance, normal
     if (endLeft) extrude._sub(normal.perp()._mult(endLeft));
     this.e3 = this.addLineVertex(currentVertex, extrude, tx, 0, endLeft, distance) - group.vertexStartIndex;
     if (this.e1 >= 0 && this.e2 >= 0) {
-        this.arrays.lineElement.emplaceBack(this.e1, this.e2, this.e3);
+        this.arrays.line.layout.element.emplaceBack(this.e1, this.e2, this.e3);
         group.elementLength++;
     }
     this.e1 = this.e2;
@@ -385,7 +385,7 @@ LineBucket.prototype.addCurrentVertex = function(currentVertex, distance, normal
     if (endRight) extrude._sub(normal.perp()._mult(endRight));
     this.e3 = this.addLineVertex(currentVertex, extrude, tx, 1, -endRight, distance) - group.vertexStartIndex;
     if (this.e1 >= 0 && this.e2 >= 0) {
-        this.arrays.lineElement.emplaceBack(this.e1, this.e2, this.e3);
+        this.arrays.line.layout.element.emplaceBack(this.e1, this.e2, this.e3);
         group.elementLength++;
     }
     this.e1 = this.e2;
@@ -420,7 +420,7 @@ LineBucket.prototype.addPieSliceVertex = function(currentVertex, distance, extru
     group.vertexLength++;
 
     if (this.e1 >= 0 && this.e2 >= 0) {
-        this.arrays.lineElement.emplaceBack(this.e1, this.e2, this.e3);
+        this.arrays.line.layout.element.emplaceBack(this.e1, this.e2, this.e3);
         group.elementLength++;
     }