--- a/js/render/draw_circle.js
+++ b/js/render/draw_circle.js
@@ -51,7 +51,6 @@ function drawCircles(painter, source, layer, coords) {
             var count = group.elementLength * 3;
             group.vaos[layer.id].bind(gl, program, vertexBuffer, paintVertexBuffer, group.vertexStartIndex, elementBuffer);
             gl.drawElements(gl.TRIANGLES, count, gl.UNSIGNED_SHORT, group.elementOffset);
-            group.vaos[layer.id].unbind(gl);
         }
     }
 }