fixes #1361

When the extension is available, use it to record and re-apply buffer binds and attrib pointer setting. When the extension isn't available this just sets the pointers and binds the buffers each time.

This also namespaces bucket buffers/arrays with objects instead of string concatenation.

I ran the frame-duration benchmark four times for master (3a95d00) and this branch:

```
vao:        15.3 ms
master:     17.9 ms
vao:        15.5 ms
master:     18.2 ms
vao:        14.2 ms
master:     18.5 ms
vao:        15.2 ms
master:     17.4 ms

vao average:    15 ms
master average: 18 ms
```

This branch takes **83%** of the time that master takes.

:eyes: @lucaswoj